# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: ketuk-api
  version: 0.2.0
  description: |
    Backend Ketuk. Sukses JSON `{"data": ...}` kecuali auth (`data` + `tokens`) dan callback Duitku (teks `SUCCESS`).
    Schema katalog + undangan + planner sesuai handler hidup. CRUD lokasi/tamu/buku ucapan sudah dikode sesuai `docs/contracts/undangan-crud.md`.
    PATCH inviter dan caption galeri juga sudah dikode sesuai `docs/contracts/undangan-patch-gallery-inviter.md`.
    CRUD planner (patch/archive, hari, kategori, item budget, checklist) sudah dikode sesuai `docs/contracts/planner-crud.md` bagian 1; rundown sesi jam+PIC belum.
    Auth dan billing masih path-only. Vendor belum diimplementasi.
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/invitations/{id}/inviters:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    get:
      tags: [Invitation]
      operationId: listInviters
      summary: Turut mengundang (bukan tamu RSVP)
      description: "Ubah satu pengundang: `/v1/invitations/{id}/inviters/{inviterId}`."
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviterListResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Invitation]
      operationId: addInviter
      summary: Tambah yang turut mengundang
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/inviters/{inviterId}:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
      - $ref: "#/components/parameters/InviterId"
    patch:
      tags: [Invitation]
      operationId: patchInviter
      summary: Ubah turut mengundang
      description: "Body parsial. `name` jika dikirim tidak boleh kosong setelah trim. `role` kosong boleh (hapus peran). Body `{}` = no-op."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchInviterRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviterResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Invitation]
      operationId: deleteInviter
      summary: Hapus turut mengundang
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Dihapus
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/gallery:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    get:
      tags: [Invitation]
      operationId: listGallery
      summary: Daftar foto galeri
      description: "Ubah caption satu foto: `/v1/invitations/{id}/gallery/{mediaId}`."
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaListResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Invitation]
      operationId: addGallery
      summary: Upload foto galeri
      description: "`multipart/form-data` field `file` (wajib), `caption` opsional. Content-Type harus `image/*`. Maks 16 MiB."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                caption:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/gallery/{mediaId}:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
      - $ref: "#/components/parameters/MediaId"
    patch:
      tags: [Invitation]
      operationId: patchGallery
      summary: Ubah caption foto galeri
      description: "Hanya `caption`. File di `./storage` tidak disentuh. Id musik → 404 `media`. Body `{}` = no-op; `caption` kosong = hapus teks."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchGalleryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Invitation]
      operationId: deleteGallery
      summary: Hapus foto galeri
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Dihapus
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/music:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    get:
      tags: [Invitation]
      operationId: getMusic
      summary: Musik undangan
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: "`data` null jika belum ada musik"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MusicResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags: [Invitation]
      operationId: setMusic
      summary: Upload musik (mengganti yang lama)
      description: "Field `file`. Content-Type harus `audio/*`. Maks 16 MiB."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Invitation]
      operationId: deleteMusic
      summary: Hapus musik
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Dihapus, atau tidak ada musik
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/gift-methods:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
    get:
      tags: [Invitation]
      operationId: listGiftMethods
      summary: Metode amplop (bank, ewallet, duitku)
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftMethodListResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Invitation]
      operationId: addGiftMethod
      summary: Tambah metode amplop
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGiftMethodRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftMethodResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/gift-methods/{methodId}:
    delete:
      tags: [Invitation]
      operationId: deleteGiftMethod
      summary: Hapus metode amplop
      security: [{bearerAuth: []}]
      parameters:
        - $ref: "#/components/parameters/InvitationId"
        - $ref: "#/components/parameters/GiftMethodId"
      responses:
        "204":
          description: Dihapus
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/invitations/{id}/gifts:
    get:
      tags: [Invitation]
      operationId: listGifts
      summary: Riwayat amplop masuk
      description: "Kas tamu (`gift_payments`), bukan `orders` langganan SaaS."
      security: [{bearerAuth: []}]
      parameters:
        - $ref: "#/components/parameters/InvitationId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftPaymentListResponse"
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/invitations/{id}/guestbook/{entryId}:
    parameters:
      - $ref: "#/components/parameters/InvitationId"
      - $ref: "#/components/parameters/GuestbookEntryId"
    patch:
      tags: [Invitation]
      operationId: patchGuestbook
      summary: Ubah entri buku ucapan
      description: "Keduanya opsional; body `{}` = no-op. Field yang dikirim tidak boleh kosong setelah trim. `created_at` tidak berubah."
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchGuestbookRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestbookResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Invitation]
      operationId: deleteGuestbook
      summary: Hapus entri buku ucapan
      security: [{bearerAuth: []}]
      responses:
        "204":
          description: Dihapus
        "400":
          $ref: "#/components/responses/InvalidId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    InvitationId:
      name: id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
    InviterId:
      name: inviterId
      in: path
      required: true
      schema: {type: string, format: uuid}
    MediaId:
      name: mediaId
      in: path
      required: true
      schema: {type: string, format: uuid}
    GiftMethodId:
      name: methodId
      in: path
      required: true
      schema: {type: string, format: uuid}
    GuestbookEntryId:
      name: entryId
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  responses:
    BadRequest:
      description: "JSON rusak (`invalid_json`) atau validasi (`invalid_input` / `invalid_id`)"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    InvalidId:
      description: Path UUID rusak
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error: {code: invalid_id, message: invalid id}
    Unauthorized:
      description: JWT hilang atau tidak valid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error: {code: unauthorized, message: missing bearer token}
...
    NotFound:
      description: Resource tidak ada atau bukan milik host; RSVP token/slug unpublished
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          examples:
            invitation:
              value:
                error: {code: not_found, message: invitation not found}
            plannerChild:
              summary: "Anak planner: day / session / category / budget_item / checklist"
              value:
                error: {code: not_found, message: budget_item not found}
...
    InternalError:
      description: Kesalahan server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
          example:
            error: {code: internal, message: internal server error}
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
...
    Inviter:
      type: object
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        role: {type: string}
        sort_order: {type: integer}
    Media:
      type: object
      properties:
        id: {type: string, format: uuid}
        kind:
          type: string
          enum: [gallery, music]
        original_name: {type: string}
        content_type: {type: string}
        url:
          type: string
          example: /v1/public/media/00000000-0000-0000-0000-000000000001
          description: Path relatif, bukan URL absolut
        caption: {type: string}
        sort_order: {type: integer}
    GuestbookEntry:
      type: object
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        message: {type: string}
        created_at: {type: string, format: date-time}
    GiftMethod:
      type: object
      properties:
        id: {type: string, format: uuid}
        kind:
          type: string
          enum: [bank, ewallet, duitku]
        label: {type: string}
        account_name: {type: string}
        account_number: {type: string}
        is_active: {type: boolean}
        sort_order: {type: integer}
    GiftPayment:
      type: object
      properties:
        id: {type: string, format: uuid}
        method_id: {type: string, format: uuid}
        sender_name: {type: string}
        message: {type: string}
        amount_idr:
          type: integer
          format: int64
        status:
          type: string
          enum: [pending, paid, failed, expired]
        merchant_order_id: {type: string}
        duitku_reference: {type: string}
        payment_url: {type: string}
        paid_at:
          type: string
          format: date-time
          nullable: true
        created_at: {type: string, format: date-time}
...
    PatchGuestbookRequest:
      type: object
      properties:
        name: {type: string, maxLength: 80}
        message: {type: string, maxLength: 2000}
    PatchInviterRequest:
      type: object
      description: Semua opsional. name jika dikirim tidak boleh kosong setelah trim. role kosong boleh.
      properties:
        name: {type: string}
        role: {type: string}
        sort_order: {type: integer}
    PatchGalleryRequest:
      type: object
      description: Hanya caption. Omit = tidak diubah. String kosong = hapus caption. Jangan kirim file.
      properties:
        caption: {type: string}
    CreateInviterRequest:
      type: object
      required: [name]
      properties:
        name: {type: string}
        role: {type: string}
        sort_order: {type: integer}
    CreateGiftMethodRequest:
      type: object
      required: [kind, label]
      properties:
        kind:
          type: string
          enum: [bank, ewallet, duitku]
        label: {type: string}
        account_name: {type: string}
        account_number: {type: string}
        sort_order: {type: integer}
...
    InviterResponse:
      type: object
      required: [data]
      properties:
        data: {$ref: "#/components/schemas/Inviter"}
    InviterListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/Inviter"}
    MediaResponse:
      type: object
      required: [data]
      properties:
        data: {$ref: "#/components/schemas/Media"}
    MediaListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/Media"}
    MusicResponse:
      type: object
      required: [data]
      properties:
        data:
          allOf:
            - $ref: "#/components/schemas/Media"
          nullable: true
    GuestbookResponse:
      type: object
      required: [data]
      properties:
        data: {$ref: "#/components/schemas/GuestbookEntry"}
...
    GiftMethodResponse:
      type: object
      required: [data]
      properties:
        data: {$ref: "#/components/schemas/GiftMethod"}
    GiftMethodListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/GiftMethod"}
...
    GiftPaymentListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: {$ref: "#/components/schemas/GiftPayment"}
...